Implement local edge reconciliation and persistence logic
- Added `current_edges.rs` to handle reconciliation of local edges for paths, including loading current targets, source facts, and applying edge replacements.
- Introduced `current_state.rs` for persisting prepared materialization results, including handling artefact and edge deletions and insertions.
- Created `local_resolution.rs` to resolve prepared local edges, extracting source facts and applying resolutions based on current targets.
- Updated tests to reflect changes in the module structure, ensuring proper loading of current edges for reconciliation.
